adjust modifiers, reorder functions, use download/build functions now
